rename: 코멘트 삭제 api 이름 수정
