+++ src/components/Board/Comment.tsx
@@ -1,7 +1,7 @@
 import ReplyImage from '@/assets/images/ic_reply_comment.svg';
 import MenuImage from '@/assets/images/ic_comment_delete.svg';
 import * as S from '@/styles/board/Comment.styled';
-import deleteComment from '@/api/deletComment';
+import deleteComment from '@/api/deleteComment';
 import { useState } from 'react';
 import formatDateTime from '@/hooks/formatDateTime';
 import useGetUserName from '@/hooks/useGetUserName';
